Extract GraphLoadingNode to separate file
