+++ Sources/PackageGraph/GraphLoadingNode.swift
@@ -1,0 +1,89 @@
+/*
+ This source file is part of the Swift.org open source project
+
+ Copyright (c) 2014 - 2020 Apple Inc. and the Swift project authors
+ Licensed under Apache License v2.0 with Runtime Library Exception
+
+ See http://swift.org/LICENSE.txt for license information
+ See http://swift.org/CONTRIBUTORS.txt for Swift project authors
+ */
+
+import TSCBasic
+import PackageLoading
+import PackageModel
+import TSCUtility
+
+/// A node used while loading the packages in a resolved graph.
+///
+/// This node uses the product filter that was already finalized during resolution.
+///
+/// - SeeAlso: DependencyResolutionNode
+public struct GraphLoadingNode: Equatable, Hashable, CustomStringConvertible {
+
+    /// The package manifest.
+    public let manifest: Manifest
+
+    /// The product filter applied to the package.
+    public let productFilter: ProductFilter
+
+    public init(manifest: Manifest, productFilter: ProductFilter) {
+        self.manifest = manifest
+        self.productFilter = productFilter
+    }
+
+    /// Returns the dependencies required by this node.
+    internal func requiredDependencies() -> [FilteredDependencyDescription] {
+        return manifest.dependenciesRequired(for: productFilter)
+    }
+
+    public var description: String {
+        switch productFilter {
+        case .everything:
+            return manifest.name
+        case .specific(let set):
+            return "\(manifest.name)[\(set.sorted().joined(separator: ", "))]"
+        }
+    }
+}
+
+/// Finds the first cycle encountered in a graph.
+///
+/// This is different from the one in tools support core, in that it handles equality separately from node traversal. Nodes traverse product filters, but only the manifests must be equal for there to be a cycle.
+internal func findCycle(
+    _ nodes: [GraphLoadingNode],
+    successors: (GraphLoadingNode) throws -> [GraphLoadingNode]
+) rethrows -> (path: [Manifest], cycle: [Manifest])? {
+    // Ordered set to hold the current traversed path.
+    var path = OrderedSet<Manifest>()
+
+    // Function to visit nodes recursively.
+    // FIXME: Convert to stack.
+    func visit(
+      _ node: GraphLoadingNode,
+      _ successors: (GraphLoadingNode) throws -> [GraphLoadingNode]
+    ) rethrows -> (path: [Manifest], cycle: [Manifest])? {
+        // If this node is already in the current path then we have found a cycle.
+        if !path.append(node.manifest) {
+            let index = path.firstIndex(of: node.manifest)!
+            return (Array(path[path.startIndex..<index]), Array(path[index..<path.endIndex]))
+        }
+
+        for succ in try successors(node) {
+            if let cycle = try visit(succ, successors) {
+                return cycle
+            }
+        }
+        // No cycle found for this node, remove it from the path.
+        let item = path.removeLast()
+        assert(item == node.manifest)
+        return nil
+    }
+
+    for node in nodes {
+        if let cycle = try visit(node, successors) {
+            return cycle
+        }
+    }
+    // Couldn't find any cycle in the graph.
+    return nil
+}
